se cambia formato de los precios por integer para integracion con mercadopago

+++ database/migrations/2021_07_03_140357_create_products_table.php
@@ -19,7 +19,7 @@ class CreateProductsTable extends Migration
             $table->string('name');
             $table->string('slug');
             $table->text('description');
-            $table->float('price');
+            $table->integer('price');
             $table->integer('quantity')->nullable();
             $table->enum('status', [Product::BORRADOR, Product::PUBLICADO])->default(Product::BORRADOR);
             $table->unsignedBigInteger('subcategory_id');
